test: 💍 refactor ssr tests and create fixtures
