+++ src/Atoms/Media/Media.test.tsx
@@ -116,7 +116,6 @@ test('Hydration', async () => {
     </ThemeProvider>
   );
   const html = renderToString(<Element />);
-  console.log(html);
   const main = document.createElement('main');
 
   main.innerHTML = html;
